refactor: merge and minimize packages
