+++ internal/backend/encode_path.go
@@ -1,4 +1,4 @@
-package system
+package backend
 
 import (
 	"encoding/base64"
