+++ internal/backend/filetree.go
@@ -1,4 +1,4 @@
-package system
+package backend
 
 import (
 	"os"
