+++ internal/backend/memlimit.go
@@ -1,4 +1,4 @@
-package memlimit
+package backend
 
 import (
 	"time"
