+++ internal/backend/shared.go
@@ -1,4 +1,4 @@
-package system
+package backend
 
 import (
 	"context"
