--- internal/web/api/download_types.go
@@ -1,13 +1,0 @@
-//go:build linux
-
-package api
-
-type VersionResponse struct {
-	Version string `json:"version"`
-}
-
-type ScriptConfig struct {
-	AgentUrl       string
-	ServerUrl      string
-	BootstrapToken string
-}
