--- internal/web/api/exclusion_types.go
@@ -1,18 +1,0 @@
-//go:build linux
-
-package api
-
-import "github.com/pbs-plus/pbs-plus/internal/store/database"
-
-type ExclusionsResponse struct {
-	Data   []database.Exclusion `json:"data"`
-	Digest string               `json:"digest"`
-}
-
-type ExclusionConfigResponse struct {
-	Errors  map[string]string   `json:"errors"`
-	Message string              `json:"message"`
-	Data    *database.Exclusion `json:"data"`
-	Status  int                 `json:"status"`
-	Success bool                `json:"success"`
-}
